added interactive guesses for 1D in notebooks
