+++ symfit/__init__.py
@@ -1,4 +1,4 @@
 from symfit.api import *
 import pkg_resources
 
-__version__  = pkg_resources.get_distribution('symfit').version
+__version__ = 1  #pkg_resources.get_distribution('symfit').version
